+++ nodeapp/core/libs/jansson/value.c
@@ -845,7 +845,7 @@ static json_t *json_real_copy(json_t *real)
 double json_number_value(const json_t *json)
 {
     if(json_is_integer(json))
-        return (double) json_integer_value(json);
+        return (double)(long)json_integer_value(json);
     else if(json_is_real(json))
         return json_real_value(json);
     else
